[47_doc_update], issue - #47, Put all the decorators into _decorators.py module and refactored all the code accordingly
