+++ lbh15/properties/lbe_properties.py
@@ -5,7 +5,7 @@ from typing import Union
 import numpy as np
 from scipy.constants import atm
 from .interface import PropertyInterface
-from .interface import range_warning
+from .._decorators import range_warning
 from .._decorators import typecheck_for_method
 from .._commons import LBE_MELTING_TEMPERATURE as T_m0
 from .._commons import LBE_BOILING_TEMPERATURE as T_b0
