+++ lbh15/properties/lbe_thermochemical_properties/diffusivity_in_lbe.py
@@ -4,7 +4,7 @@ from typing import List
 import numpy as np
 from scipy.constants import atm, R
 from lbh15.properties.interface import PropertyInterface
-from lbh15.properties.interface import range_warning
+from ..._decorators import range_warning
 from ..._decorators import typecheck_for_method
