+++ lbh15/properties/lbe_thermochemical_properties/lbe_oxygen_limits.py
@@ -6,7 +6,6 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from ..interface import PropertyInterface
-from ..interface import range_warning
 from .solubility_in_lbe import OxygenSolubility
 from .solubility_in_lbe import ChromiumSolubilityGosse2014
 from .solubility_in_lbe import ChromiumSolubilityCourouau2004
@@ -16,6 +15,7 @@ from .solubility_in_lbe import NickelSolubilityMartinelli2010
 from .solubility_in_lbe import IronSolubilityGosse2014
 from .solubility_in_lbe import IronSolubilityWeeks1969
 from .lbe_thermochemical import LeadChemicalActivity
+from ..._decorators import range_warning
 from ..._decorators import typecheck_for_method
