+++ lbh15/properties/lbe_thermochemical_properties/solubility_in_lbe.py
@@ -5,7 +5,7 @@ from typing import Union
 import numpy as np
 from scipy.constants import atm
 from lbh15.properties.interface import PropertyInterface
-from lbh15.properties.interface import range_warning
+from ..._decorators import range_warning
 from ..._decorators import typecheck_for_method
