+++ lbh15/properties/lead_thermochemical_properties/diffusivity_in_lead.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from lbh15.properties.interface import PropertyInterface
-from lbh15.properties.interface import range_warning
+from ..._decorators import range_warning
 from ..._decorators import typecheck_for_method
