+++ lbh15/properties/lead_thermochemical_properties/lead_thermochemical.py
@@ -6,11 +6,11 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from lbh15.properties.interface import PropertyInterface
-from lbh15.properties.interface import range_warning
 from ..lead_properties import h
 from ..._commons import LEAD_MELTING_TEMPERATURE as T_m0
 from ..._commons import LEAD_MOLAR_MASS as M
 from ..._commons import OXYGEN_MOLAR_MASS as M_O
+from ..._decorators import range_warning
 from ..._decorators import typecheck_for_method
